* feat(sdk): migrate changes to prebid mobile

+++ PrebidMobile/PrebidMobileRendering/ImpressionTask.swift
@@ -18,5 +18,5 @@ import Foundation
 
 struct ImpressionTask {
     let task: (_ completion: @escaping (() -> Void)) -> Void
-    let delayInterval: Double
+    let delayInterval: Int
 }
